250205 추천/비추천 리팩토링중 게시글 추천/비추천, 댓글 추천/비추천 테이블 2개로 분리

+++ src/main/java/com/jhm/springbootwebservice/web/dto/request/RecommendRequestDto.java
@@ -12,7 +12,7 @@ public class RecommendRequestDto {
     private Long postId;
     private Long userId;
     private Long commentId;
-    private Long recommendType;
+    private int recommendType;
 
     public RecommendRequestDto(Long postId, Long userId) {
         this.postId = postId;
